Moved skeleton buff and list command to own package
