+++ src/main/java/me/glaremasters/guilds/commands/CommandGuilds.java
@@ -84,20 +84,6 @@ public class CommandGuilds extends BaseCommand {
     private Economy economy;
 
     /**
-     * List all the guilds on the server
-     * @param player the player executing this command
-     */
-    @Subcommand("list")
-    @Description("{@@descriptions.list}")
-    @CommandPermission(Constants.BASE_PERM + "list")
-    public void onGuildList(Player player) {
-        //todo after explanation waiting for @Glare
-        playerPages.put(player.getUniqueId(), 1);
-        // guildList = getSkullsPage(1);
-        player.openInventory(guildList);
-    }
-
-    /**
      * Request an invite
      *
      * @param player the player requesting
@@ -131,184 +117,6 @@ public class CommandGuilds extends BaseCommand {
             }
         }
         getCurrentCommandIssuer().sendInfo(Messages.REQUEST__SUCCESS, "{guild}", targetGuild.getName());
-    }
-
-    /**
-     * Open the guild buff menu
-     * @param player the player opening the menu
-     * @param guild the guild which's player is opening the menu
-     * @param role the role of the player
-     */
-    @Subcommand("buff")
-    @Description("{@@descriptions.buff}")
-    @CommandPermission(Constants.BASE_PERM + "buff")
-    public void onBuff(Player player, Guild guild, GuildRole role) {
-        if (!role.isActivateBuff()) {
-            getCurrentCommandIssuer().sendInfo(Messages.ERROR__ROLE_NO_PERMISSION);
-            return;
-        }
-        Inventory buff = Bukkit.createInventory(null, 9, settingsManager.getProperty(GuiSettings.GUILD_BUFF_NAME));
-        List<String> lore = new ArrayList<>();
-/*        createBuffItem("haste", lore, buff, 0);
-        createBuffItem("speed", lore, buff, 1);
-        createBuffItem("fire-resistance", lore, buff, 2);
-        createBuffItem("night-vision", lore, buff, 3);
-        createBuffItem("invisibility", lore, buff, 4);
-        createBuffItem("strength", lore, buff, 5);
-        createBuffItem("jump", lore, buff, 6);
-        createBuffItem("water-breathing", lore, buff, 7);
-        createBuffItem("regeneration", lore, buff, 8);*/
-        player.openInventory(buff);
-    }
-
-    /**
-     * Create an item stack for the list
-     * @param mat
-     * @param name
-     * @param lore
-     * @return
-     */
-    private ItemStack createItemStack(Material mat, String name, List<String> lore) {
-        ItemStack itemStack = new ItemStack(mat);
-
-        ItemMeta meta = itemStack.getItemMeta();
-        meta.setDisplayName(name);
-        meta.setLore(lore);
-
-        itemStack.setItemMeta(meta);
-        return itemStack;
-    }
-
-    //todo rewrite this + explanation plz @Glare
-//    /**
-//     * Create an item for buff
-//     * @param buffName
-//     * @param name
-//     * @param buff
-//     * @param slot
-//     */
-//    private void createBuffItem(String buffName, List<String> name, Inventory buff, int slot) {
-//        getStringList("buff.description." + buffName).stream().map(ConfigUtils::StringUtils.color).forEach(name::add);
-//        name.add("");
-//        name.add(getString("buff.description.price") + getString("buff.price." + buffName));
-//        name.add(getString("buff.description.length") + getString("buff.time." + buffName));
-//        if (getBoolean("buff.display." + buffName)) {
-//            buff.setItem(slot, createItemStack(Material.getMaterial(getString("buff.icon." + buffName)), getString("buff.name." + buffName), name));
-//        }
-//        name.clear();
-//    }
-//
-//    /**
-//     * Handling for the list page
-//     * @param page
-//     * @return
-//     */
-//    public static Inventory getSkullsPage(int page) {
-//        Map<UUID, ItemStack> skulls = new HashMap<>();
-//        Inventory inv = Bukkit.createInventory(null, 54, getString("guild-list.gui-name"));
-//
-//        int startIndex = 0;
-//        int endIndex = 0;
-//
-//        Guilds.getGuilds().getGuildHandler().getGuilds().values().forEach(guild -> {
-//            ItemStack skull = HeadUtils.getSkull(HeadUtils.getTextureUrl(guild.getGuildMaster().getUniqueId()));
-//            SkullMeta skullMeta = (SkullMeta) skull.getItemMeta();
-//            List<String> lore = new ArrayList<>();
-//
-//            getStringList("guild-list.head-lore").forEach(line -> lore.add(StringUtils.color(line)
-//                    .replace("{guild-name}", guild.getName())
-//                    .replace("{guild-prefix}", guild.getPrefix())
-//                    .replace("{guild-master}", Bukkit.getOfflinePlayer(guild.getGuildMaster().getUniqueId()).getName())
-//                    .replace("{guild-status}", guild.getStatus())
-//                    .replace("{guild-tier}", String.valueOf(guild.getTier()))
-//                    .replace("{guild-balance}", String.valueOf(guild.getBalance()))
-//                    .replace("{guild-member-count}", String.valueOf(guild.getMembers().size()))));
-//
-//            skullMeta.setLore(lore);
-//
-//            String name = Bukkit.getOfflinePlayer(guild.getGuildMaster().getUniqueId()).getName();
-//            skullMeta.setDisplayName(getString("guild-list.item-name").replace("{player}", name).replace("{guild-name}", guild.getName()));
-//            skull.setItemMeta(skullMeta);
-//            skulls.put(guild.getGuildMaster().getUniqueId(), skull);
-//        });
-//
-//        ItemStack previous = new ItemStack(Material.getMaterial(getString("guild-list.previous-page-item")), 1);
-//        ItemMeta previousMeta = previous.getItemMeta();
-//        previousMeta.setDisplayName(getString("guild-list.previous-page-item-name"));
-//        previous.setItemMeta(previousMeta);
-//        ItemStack next = new ItemStack(Material.getMaterial(getString("guild-list.next-page-item")), 1);
-//        ItemMeta nextMeta = next.getItemMeta();
-//        nextMeta.setDisplayName(getString("guild-list.next-page-item-name"));
-//        next.setItemMeta(nextMeta);
-//        ItemStack barrier = new ItemStack(Material.getMaterial(getString("guild-list.page-number-item")), 1);
-//        ItemMeta barrierMeta = barrier.getItemMeta();
-//        barrierMeta.setDisplayName(getString("guild-list.page-number-item-name").replace("{page}", String.valueOf(page)));
-//        barrier.setItemMeta(barrierMeta);
-//        inv.setItem(53, next);
-//        inv.setItem(49, barrier);
-//        inv.setItem(45, previous);
-//
-//        startIndex = (page - 1) * 45;
-//        endIndex = startIndex + 45;
-//
-//        if (endIndex > skulls.values().size()) {
-//            endIndex = skulls.values().size();
-//        }
-//
-//        int iCount = 0;
-//        for (int i1 = startIndex; i1 < endIndex; i1++) {
-//            inv.setItem(iCount, (ItemStack) skulls.values().toArray()[i1]);
-//            iCount++;
-//        }
-//
-//        return inv;
-//    }
-//
-//    /**
-//     * Create player skull
-//     * @param player
-//     * @return
-//     */
-//    public ItemStack createSkull(Player player) {
-//        ItemStack skull = new ItemStack(Material.SKULL_ITEM, 1, (short) SkullType.PLAYER.ordinal());
-//
-//        SkullMeta meta = (SkullMeta) skull.getItemMeta();
-//        meta.setOwner(player.getName());
-//        meta.setDisplayName(getString("info.playername").replace("{player-name}", player.getName()));
-//
-//        List<String> info = new ArrayList<>();
-//        info.add(getString("info.kills").replace("{kills}", String.valueOf(player.getStatistic(Statistic.PLAYER_KILLS))));
-//        info.add(getString("info.deaths").replace("{deaths}", String.valueOf(player.getStatistic(Statistic.DEATHS))));
-//        meta.setLore(info);
-//
-//        skull.setItemMeta(meta);
-//        return skull;
-//    }
-//
-//
-    /**
-     * Checks the name requirements from the config.
-     * @param name the name to check
-     * @return a boolean if the name is wrong
-     */
-    private boolean nameMeetsRequirements(String name) {
-        String regex = settingsManager.getProperty(GuildSettings.NAME_REQUIREMENTS);
-
-        if (!name.matches(regex)) {
-            getCurrentCommandIssuer().sendInfo(Messages.CREATE__REQUIREMENTS);
-            return true;
-        }
-
-        if (settingsManager.getProperty(GuildSettings.BLACKLIST_TOGGLE)) {
-            for (String word : settingsManager.getProperty(GuildSettings.BLACKLIST_WORDS)) {
-                if (name.toLowerCase().contains(word)) {
-                    getCurrentCommandIssuer().sendInfo(Messages.ERROR__BLACKLIST);
-                    return true;
-                }
-            }
-        }
-
-        return false;
     }
 
 }
